pyproject.toml
[project]
name = "openwhale"
version = "0.1.0"
description = "Read the radme silly"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.6.4",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "ollama>=0.6.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pylint>=4.0.4",
]
